refactor: delete handlers.py shim + move state-axis tests + extract Redirects
- Delete `src/framework/dispatch/handlers.py` (the re-export shim from
  PR #1088). Factory infrastructure (_FactoryShape, shape constants,
  _make_factory_handler) moves to `mounts/_factory.py`; the 7 lazy
  imports in the mount files are updated accordingly.
- Update the two domain test files that still imported through the shim
  (`users/test_handlers.py`, `clinicians/test_handlers.py`) to import
  directly from `mounts.detail`, `mounts.create`, `mounts.list_`.
- Move the 2 state-axis self-guard tests from `test_handlers.py` into
  `mounts/test_state_axis.py` (colocated with the mount they cover),
  then delete `test_handlers.py`.
- Extract `Redirects` from `entity_spec.py` into `redirects.py`; add
  a re-export in `entity_spec.py` for backward compat; update the 5
  callers (4 spec files + test_entity_spec.py) to import from
  `redirects` directly.
- Update `dispatch/README.md` to reflect the removed shim, new
  `_factory.py` home, `redirects.py` extraction, and relocated tests.
