+++ src/domain/specs/program.py
@@ -20,9 +20,9 @@ from src.framework.dispatch.entity_spec import (
     AUTHENTICATED,
     OWNER_OR_ADMIN,
     EntitySpec,
-    Redirects,
     RouteSet,
 )
+from src.framework.dispatch.redirects import Redirects
 
 _program_form_redirect = Redirects.to_edit_form("programs", "program_id")
